refactor(replication)!: collapse state sync to a single updated protocol version
The pre-existing state sync wire protocol never shipped, so there is no compatibility story to keep: version 1 simply becomes the updated protocol (atomic restore, indexed-tree transfer, PrivateDocumentStore replay, aggregate-restore fix). Remove the dual-version machinery: SUPPORTED_STATE_SYNC_VERSIONS, is_supported_state_sync_version, INDEXED_SYNC_MIN_VERSION, and the version-gated indexed-tree serving/discovery rejects. Every entry point now compares the requested version against CURRENT_STATE_SYNC_VERSION (back to 1) and rejects anything else with an error naming both versions; the version parameter threading and the session-consistency check stay as the hook for a future bump.

Replace the cross-version compat tests with replication_version_tests.rs pinning what survives: descriptive rejection of unsupported versions on both sides with an untouched destination, the session-version-consistency guard, and a round-trip sanity anchor. Reword the append-only page-cursor check as plain input validation.
