+++ grovedb/src/tests/mod.rs
@@ -99,11 +99,11 @@ mod read_mode_gate_tests;
 mod reference_path_tests;
 mod reference_with_sum_item_tests;
 mod replication_checkpoint_prune_tests;
-mod replication_compat_tests;
 mod replication_fuzz_tests;
 mod replication_scale_tests;
 mod replication_session_tests;
 mod replication_utils_tests;
+mod replication_version_tests;
 mod run_path_query_tests;
 mod snapshot_read_transaction_tests;
 mod succinctness_gap_test;
